Create a weighted protein-protein interaction network based on cell-type specific correlation coefficients.

Create a graph based on gene expression correlation for the protein-protein interactions for each population. The edges are the intersection of the ones in the protein-protein interaction network and the ones obtained from the correlation of gene expression in the cells, the weights are the correlation value. The edge lists are saved to files. A more efficient version of cellnetworksPPI.

Usage

cellnetworksPPI2(
  ppi,
  expr,
  samples,
  corThr1,
  corThr2,
  column,
  dir.prefix = "results/paths"
)

Arguments

ppi

igraph graph; the network as calculated by the createPPI function.

expr

dgCMatrix; gene expression in each cell, found in CellRouter@ndata.

samples

data frame; the metadata information of the cells, found in CellRouter@sampTab.

corThr1

numeric; values smaller then corThr1 in the correlation matrix will be replaced by 0. Therefore, there will be no connections between the genes with value 0.

corThr2

numeric; keep gene-gene interactions with correlation higher than the threshold.

column

character; character; column in the metadata table to specify how the data should be aggregated (e.g. per cell type, cluster or other metadata available).

dir.prefix

character; prefix of the directory where the files will be saved.

Value

list; the networks (igraph objects) for each cell type/cluster/other metadata.