Plot boxplot with gene signatures.

Plot boxplot of gene expression for the specified genes in each population.

Usage

plotSignaturesBoxplot(
  object,
  assay.type = "RNA",
  markers,
  column.ann,
  column.color,
  num.cells = NULL,
  threshold = 2,
  genes.show = NULL,
  low = "purple",
  intermediate = "black",
  high = "yellow",
  order = NULL,
  fontsize = 5
)

Arguments

object

CellRouter object.

assay.type

character; the type of data to use.

markers

tibble; genes preferentially expressed in each column.ann, as obtained in the findSignatures function. For example, in clusters or sorted populations.

column.ann

character; column in the metadata table used to annotate the kNN graph. For example, clusters, sorted cell populations.

column.color

character; column corresponding to the colors.

num.cells

numeric; number of cells to show in the heatmap.

threshold

numeric; threshold used to center the data.

genes.show

character vector; gene names to show in the heatmap. The default is the top 5 genes according to fold change for each population.

low

character; color for low expression.

intermediate

character; color for intermediate expression.

high

character; color for high expression.

order

character vector; order of population names, the default is alphabetical order.

fontsize

numeric; font size.

Value

matrix; annotations.